Move atomics (and other threading stuff)
... into a dedicated file.

None of the C code has changed - just been moved
