+++ Cython/Utility/MemoryView_C.c
@@ -16,163 +16,6 @@ typedef struct {
 #define __Pyx_MemoryView_Len(m)  (m.shape[0])
 
 
-/////////// Atomics.proto /////////////
-//@proto_block: utility_code_proto_before_types
-
-#include <pythread.h>
-
-#ifndef CYTHON_ATOMICS
-    #define CYTHON_ATOMICS 1
-#endif
-// using CYTHON_ATOMICS as a cdef extern bint in the Cython memoryview code
-// interacts badly with "import *". Therefore, define a helper function-like macro
-#define __PYX_CYTHON_ATOMICS_ENABLED() CYTHON_ATOMICS
-#define __PYX_GET_CYTHON_COMPILING_IN_CPYTHON_FREETHREADING() CYTHON_COMPILING_IN_CPYTHON_FREETHREADING
-
-#define __pyx_atomic_int_type int
-#define __pyx_nonatomic_int_type int
-
-// For standard C/C++ atomics, get the headers first so we have ATOMIC_INT_LOCK_FREE
-// defined when we decide to use them.
-#if CYTHON_ATOMICS && (defined(__STDC_VERSION__) && \
-                        (__STDC_VERSION__ >= 201112L) && \
-                        !defined(__STDC_NO_ATOMICS__))
-    #include <stdatomic.h>
-#elif CYTHON_ATOMICS && (defined(__cplusplus) && ( \
-                    (__cplusplus >= 201103L) || \
-                    (defined(_MSC_VER) && _MSC_VER >= 1700)))
-    #include <atomic>
-#endif
-
-#if CYTHON_ATOMICS && (defined(__STDC_VERSION__) && \
-                        (__STDC_VERSION__ >= 201112L) && \
-                        !defined(__STDC_NO_ATOMICS__) && \
-                       ATOMIC_INT_LOCK_FREE == 2)
-    // C11 atomics are available and  ATOMIC_INT_LOCK_FREE is definitely on
-    #undef __pyx_atomic_int_type
-    #define __pyx_atomic_int_type atomic_int
-    #define __pyx_atomic_ptr_type atomic_uintptr_t
-    #define __pyx_nonatomic_ptr_type uintptr_t
-    #define __pyx_atomic_incr_relaxed(value) atomic_fetch_add_explicit(value, 1, memory_order_relaxed)
-    #define __pyx_atomic_incr_acq_rel(value) atomic_fetch_add_explicit(value, 1, memory_order_acq_rel)
-    #define __pyx_atomic_decr_acq_rel(value) atomic_fetch_sub_explicit(value, 1, memory_order_acq_rel)
-    #define __pyx_atomic_sub(value, arg) atomic_fetch_sub(value, arg)
-    #define __pyx_atomic_int_cmp_exchange(value, expected, desired) atomic_compare_exchange_strong(value, expected, desired)
-    #define __pyx_atomic_load(value) atomic_load(value)
-    #define __pyx_atomic_store(value, new_value) atomic_store(value, new_value)
-    #define __pyx_atomic_pointer_load_relaxed(value) atomic_load_explicit(value, memory_order_relaxed)
-    #define __pyx_atomic_pointer_load_acquire(value) atomic_load_explicit(value, memory_order_acquire)
-    #define __pyx_atomic_pointer_exchange(value, new_value) atomic_exchange(value, (__pyx_nonatomic_ptr_type)new_value)
-    #if defined(__PYX_DEBUG_ATOMICS) && defined(_MSC_VER)
-        #pragma message ("Using standard C atomics")
-    #elif defined(__PYX_DEBUG_ATOMICS)
-        #warning "Using standard C atomics"
-    #endif
-#elif CYTHON_ATOMICS && (defined(__cplusplus) && ( \
-                    (__cplusplus >= 201103L) || \
-                    /*_MSC_VER 1700 is Visual Studio 2012 */ \
-                    (defined(_MSC_VER) && _MSC_VER >= 1700)) && \
-                    ATOMIC_INT_LOCK_FREE == 2)
-    // C++11 atomics are available and ATOMIC_INT_LOCK_FREE is definitely on
-    #undef __pyx_atomic_int_type
-    #define __pyx_atomic_int_type std::atomic_int
-    #define __pyx_atomic_ptr_type std::atomic_uintptr_t
-    #define __pyx_nonatomic_ptr_type uintptr_t
-    #define __pyx_atomic_incr_relaxed(value) std::atomic_fetch_add_explicit(value, 1, std::memory_order_relaxed)
-    #define __pyx_atomic_incr_acq_rel(value) std::atomic_fetch_add_explicit(value, 1, std::memory_order_acq_rel)
-    #define __pyx_atomic_decr_acq_rel(value) std::atomic_fetch_sub_explicit(value, 1, std::memory_order_acq_rel)
-    #define __pyx_atomic_sub(value, arg) std::atomic_fetch_sub(value, arg)
-    #define __pyx_atomic_int_cmp_exchange(value, expected, desired) std::atomic_compare_exchange_strong(value, expected, desired)
-    #define __pyx_atomic_load(value) std::atomic_load(value)
-    #define __pyx_atomic_store(value, new_value) std::atomic_store(value, new_value)
-    #define __pyx_atomic_pointer_load_relaxed(value) std::atomic_load_explicit(value, std::memory_order_relaxed)
-    #define __pyx_atomic_pointer_load_acquire(value) std::atomic_load_explicit(value, std::memory_order_acquire)
-    #define __pyx_atomic_pointer_exchange(value, new_value) std::atomic_exchange(value, (__pyx_nonatomic_ptr_type)new_value)
-
-    #if defined(__PYX_DEBUG_ATOMICS) && defined(_MSC_VER)
-        #pragma message ("Using standard C++ atomics")
-    #elif defined(__PYX_DEBUG_ATOMICS)
-        #warning "Using standard C++ atomics"
-    #endif
-#elif CYTHON_ATOMICS && (__GNUC__ >= 5 || (__GNUC__ == 4 && \
-                    (__GNUC_MINOR__ > 1 ||  \
-                    (__GNUC_MINOR__ == 1 && __GNUC_PATCHLEVEL__ >= 2))))
-    /* gcc >= 4.1.2 */
-    #define __pyx_atomic_ptr_type void*
-    #define __pyx_atomic_incr_relaxed(value) __sync_fetch_and_add(value, 1)
-    #define __pyx_atomic_incr_acq_rel(value) __sync_fetch_and_add(value, 1)
-    #define __pyx_atomic_decr_acq_rel(value) __sync_fetch_and_sub(value, 1)
-    #define __pyx_atomic_sub(value, arg) __sync_fetch_and_sub(value, arg)
-    static CYTHON_INLINE int __pyx_atomic_int_cmp_exchange(__pyx_atomic_int_type* value, __pyx_nonatomic_int_type* expected, __pyx_nonatomic_int_type desired) {
-        __pyx_nonatomic_int_type old = __sync_val_compare_and_swap(value, *expected, desired);
-        int result = old == *expected;
-        *expected = old;
-        return result;
-    }
-    // the legacy gcc sync builtins don't seem to have plain "load" or "store".
-    #define __pyx_atomic_load(value) __sync_fetch_and_add(value, 0)
-    #define __pyx_atomic_store(value, new_value) __sync_lock_test_and_set(value, new_value)
-    #define __pyx_atomic_pointer_load_relaxed(value) __sync_fetch_and_add(value, 0)
-    #define __pyx_atomic_pointer_load_acquire(value) __sync_fetch_and_add(value, 0)
-    #define __pyx_atomic_pointer_exchange(value, new_value) __sync_lock_test_and_set(value, (__pyx_atomic_ptr_type)new_value)
-
-    #ifdef __PYX_DEBUG_ATOMICS
-        #warning "Using GNU atomics"
-    #endif
-#elif CYTHON_ATOMICS && defined(_MSC_VER)
-    /* msvc */
-    #include <intrin.h>
-    #undef __pyx_atomic_int_type
-    #define __pyx_atomic_int_type long
-    #define __pyx_atomic_ptr_type void*
-    #undef __pyx_nonatomic_int_type
-    #define __pyx_nonatomic_int_type long
-    #pragma intrinsic (_InterlockedExchangeAdd, _InterlockedExchange, _InterlockedCompareExchange, _InterlockedCompareExchangePointer, _InterlockedExchangePointer)
-    #define __pyx_atomic_incr_relaxed(value) _InterlockedExchangeAdd(value, 1)
-    #define __pyx_atomic_incr_acq_rel(value) _InterlockedExchangeAdd(value, 1)
-    #define __pyx_atomic_decr_acq_rel(value) _InterlockedExchangeAdd(value, -1)
-    #define __pyx_atomic_sub(value, arg) _InterlockedExchangeAdd(value, -arg)
-    static CYTHON_INLINE int __pyx_atomic_int_cmp_exchange(__pyx_atomic_int_type* value, __pyx_nonatomic_int_type* expected, __pyx_nonatomic_int_type desired) {
-        __pyx_nonatomic_int_type old = _InterlockedCompareExchange(value, desired, *expected);
-        int result = old == *expected;
-        *expected = old;
-        return result;
-    }
-    #define __pyx_atomic_load(value) _InterlockedExchangeAdd(value, 0)
-    #define __pyx_atomic_store(value, new_value) _InterlockedExchange(value, new_value)
-    // Microsoft says that simple reads are guaranteed to be atomic.
-    // https://learn.microsoft.com/en-gb/windows/win32/sync/interlocked-variable-access?redirectedfrom=MSDN
-    // The volatile cast is what CPython does.
-    #define __pyx_atomic_pointer_load_relaxed(value) *(void * volatile *)value
-    // compare/exchange is probably overkill nonsense, but plain "load" intrinsics are hard to get.
-    #define __pyx_atomic_pointer_load_acquire(value) _InterlockedCompareExchangePointer(value, 0, 0)
-    #define __pyx_atomic_pointer_exchange(value, new_value) _InterlockedExchangePointer(value, (__pyx_atomic_ptr_type)new_value)
-
-    #ifdef __PYX_DEBUG_ATOMICS
-        #pragma message ("Using MSVC atomics")
-    #endif
-#else
-    #undef CYTHON_ATOMICS
-    #define CYTHON_ATOMICS 0
-
-    #ifdef __PYX_DEBUG_ATOMICS
-        #warning "Not using atomics"
-    #endif
-#endif
-
-#if CYTHON_ATOMICS
-    #define __pyx_add_acquisition_count(memview) \
-             __pyx_atomic_incr_relaxed(__pyx_get_slice_count_pointer(memview))
-    #define __pyx_sub_acquisition_count(memview) \
-            __pyx_atomic_decr_acq_rel(__pyx_get_slice_count_pointer(memview))
-#else
-    #define __pyx_add_acquisition_count(memview) \
-            __pyx_add_acquisition_count_locked(__pyx_get_slice_count_pointer(memview), memview->lock)
-    #define __pyx_sub_acquisition_count(memview) \
-            __pyx_sub_acquisition_count_locked(__pyx_get_slice_count_pointer(memview), memview->lock)
-#endif
-
-
 /////////////// ObjectToMemviewSlice.proto ///////////////
 
 static CYTHON_INLINE {{memviewslice_name}} {{funcname}}(PyObject *, int writable_flag);
